+++ front-end/src/actions/actions.js
@@ -2,17 +2,18 @@ import { axiosWithAuth, storeUser } from '../utils/utils';
 
 
 
+
 export const authIn = (userData, locationHistory) => dispatch => {
     const authType = locationHistory.location.pathname.match(/login/) ? 'login' : 'register';
     dispatch({type: 'LOADING'});
 
     axiosWithAuth().post(`/auth/${userData.role}/${authType}`, userData)
                    .then( async res => {
-                       console.log('AUTH ATTEMPT DATA: ', res);
+                    //    console.log('AUTH ATTEMPT DATA: ', res);
                         //    WORK AROUND FOR MISRETURNED DATA FROM BACKEND
                         const user = res.data.user ? res.data.user : res.data.data.user;
                        locationHistory.push( user.role === 'admin' ? '/home/admin' : `/home/user/${user.userId}` );
-                       console.log(res.data);
+                    //    console.log(res.data);
                        //Work around for mis-shaped data on return
                        res.data.token ? localStorage.setItem('token', res.data.token) : localStorage.setItem('token', res.data.user.token);
                        const {email, role, userTickets} = await res.data.user;
@@ -32,7 +33,7 @@ export const authIn = (userData, locationHistory) => dispatch => {
 export const getTickets = userID => dispatch => {
     axiosWithAuth().get(`/api/tickets/${userID}`)
                    .then(res => {
-                     console.log('*****: ', res.data)
+                     
                      const currentUser = JSON.parse( localStorage.getItem('user') );
                      localStorage.setItem('user', JSON.stringify({...currentUser, userTickets: res.data.tickets}));
                      dispatch({type: 'UPDATE_TICKETS', payload: res.data.tickets});
@@ -58,13 +59,34 @@ export const userTickets = (id) => dispatch => {
     })
 }
 
-export const addticket=(ticket,id)=>dispatch=>{
+export const addticket=(ticket,id,history)=>dispatch=>{
     axiosWithAuth().post(`/api/tickets/${id}`,ticket)
+    
     .then(res=>{
         console.log(res)
         dispatch({type:"ADD_TICKET", payload:res.data.ticket})
+        history.push(`/home/user/${id}`)
     })
     .catch(err=>{
         console.log(err)
+    })
+}
+
+
+export const updateticket=(ticket,id)=>dispatch=>{
+    axiosWithAuth().put(`/api/tickets/${id}`,ticket)
+    .then(res=>{
+        console.log(res)
+        // dispatch({type:"UPDATE_TICKET", payload:res.data.ticket})
+    })
+    .catch(err=>{
+        console.log(err)
+    })
+}
+
+export const getoneticket=(id)=>dispatch=>{
+    axiosWithAuth().get(`/api/tickets/${id}`)
+    .then(res=>{
+        console.log(res)
     })
 }
