+++ front-end/src/reducers/DevDeskReducer.js
@@ -5,7 +5,8 @@ const intialState = {
         id: '',
         role: '',
         name: '',
-        userTickets: []
+        userTickets: [],
+        oneTicket:{}
     }
 }
 
@@ -53,6 +54,13 @@ export const devDeskReducer = (state = intialState, action) => {
                 }
             }
         
+            case "GET_ONE_TICKET":
+                return{
+                    ...state,
+                    loading:false,
+                    oneTicket:action.payload
+                }
+
         case 'UPDATE_TICKETS':
                 return {...state, loading: false, user: { ...state.user, tickets: action.payload.map(ticket => ticket.ticket_id)}};
